feat(spgen): generate the action selection and its dispatch
- 585728 reachable combinations compared through spshell, 1171456 answers,
  zero disagreements, in 360ms
- the plugin gets a table walked lazily, so a predicate is asked for only
  when the walk reaches it and the side effects stay where they were
- the stranded combinations are pinned at exactly the 208 the shipped chain
  strands, so the port carries the hole rather than fixing it
- all 17 action and reason call sites carried byte for byte, asserted against
  a pinned revision of nextbot_behavior.sp
- make gen writes both files

+++ internal/spgen/lazy_test.go
@@ -45,7 +45,7 @@ func TestLazyWalkAgreesWithSelect(t *testing.T) {
 	compared := 0
 	for p := range sweep {
 		want := actionsel.Select(p.state, p.class, p.flags())
-		got, _, err := table.Walk(p.axes(), func(id int32) bool { return p.bits&(1<<uint(id)) != 0 })
+		got, _, err := table.Walk(p.axes(), func(id int32) bool { return p.bits&(1<<id) != 0 })
 		if err != nil {
 			t.Fatalf("%s: %v", p, err)
 		}
@@ -86,7 +86,7 @@ func TestTheWalkAsksNothingExtra(t *testing.T) {
 		if out != int32(actionsel.Select(p.state, p.class, p.flags())) {
 			t.Fatalf("%s: the interpreter answers %d, the compiled Go answers %d", p, out, int32(actionsel.Select(p.state, p.class, p.flags())))
 		}
-		_, asked, err := table.Walk(p.axes(), func(id int32) bool { return p.bits&(1<<uint(id)) != 0 })
+		_, asked, err := table.Walk(p.axes(), func(id int32) bool { return p.bits&(1<<id) != 0 })
 		if err != nil {
 			t.Fatalf("%s: %v", p, err)
 		}
